+++ config/initializers/session_store.rb
@@ -3,4 +3,4 @@
 # Be sure to restart your server when you modify this file.
 
 # rubocop:disable Metrics/LineLength
-Rails.application.config.session_store :cookie_store, key: '_vesta_session', domain: ".#{ENV.fetch('APPLICATION_HOST').gsub(/:\d+/, '')}"
+Rails.application.config.session_store :cookie_store, key: '_vesta_session', expire_after: 1.year, domain: ".#{ENV.fetch('APPLICATION_HOST').gsub(/:\d+/, '')}"